Working, for some definition of working.
Currently printing out what would be passed to the BatchJob() init.
Waiting for pbs on donkey and tweaks to the job_runner interface.
Time to work on refining the tool sample XMLs to break them up,
as a result of our decision to have a <tool> run as a single
Torque job.
Also need to finish up the validate stuff.